Make coronahack.py compatible with typing system
Declare the types of inputs and outputs of pipeline steps in the
'coronahack.py' script. This is needed for our KF 1.5 release which
introduces Kale's typing system.

Also, reduce the number of epochs from 10 to 4 since after the fourth
epoch, the model does not improve, so the whole process is stalled.

The new version of the 'coronahack.py' is backwards compatible with our
1.4 release.

Refs arrikto/dev#1738
Refs arrikto/dev#1502
Refs arrikto/dev#1952

Signed-off-by: Kostis Andriopoulos <kandrio@arrikto.com>
Reviewed-by: Stefano Fioravanzo <stefano@arrikto.com>