Revert TF 2.6.2 version bump and dependency changes
The recent version bump to TF 2.6.2 and Python 3.8.0 brought along
changes to the dependencies of some of our examples.

In particular, the pin `pandas==1.4.2` in the OpenVaccine example
broke it when attempting to run in containers using Python 3.6.

Revert the recent changes to keep on supporting Python 3.6 and
Tensorflow 2.3.0.

This commit reverts the following commits:
1. 707a93e724b34ab7dc0777d66960d975f5cd2701
2. b931efb224f5084f65a007126775aa1b1d1e227f
3. 5d9672000dcfcc16434700ed1c83a89d4a1b728f
4. af3727c38d046bf528388c9cfe23295dffe8706b

Refs arrikto/dev#2032

Signed-off-by: George Alexopoulos <grgalex@arrikto.com>
Reviewed-by: Dimitris Poulopoulos <dimpo@arrikto.com>