coronahack: Reconcile Coronack with the new PyTorch marshal backend
Introduce changes to the Coronahack example due to the new PyTorch
marshaling backend.

The PyTorch marshaling backend creates an archive that includes the
serialized model's weights and its definition (i.e., the Python module). Thus,
it expects the definition of the model to reside in a different Python module,
than the one defining the KFP steps so that it can reference it during the
serialization process.

When it loads the model, it first imports and executes the Python module
to create the model object and then loads the weights.

Refs arrikto/dev#1752
Refs arrikto/dev#1889

Signed-off-by: Dimitris Poulopoulos <dimpo@arrikto.com>
Reviewed-by: George Alexopoulos <grgalex@arrikto.com>