coronahack: Change torchmetrics requirement
Change CoronaHack's torchmetrics requirement, to take into account the
extra requirements needed when working with text.

Specifically, torchmetrics needs a version of the tqdm library that is
not included in the default requirements. Thus, it is necessary to install
the extra requirements with the name 'text' of the torchmetrics library.

Closes arrikto/dev#2220

Signed-off-by: Dorothea Kalliora <dorothea@arrikto.com>
Reviewed-by: Dimitris Poulopoulos <dimpo@arrikto.com>