bulldozers: Make the Bulldozers example run faster
Make the bulldozers example run faster since the new version of
AutoSklearn required generates several configurations that explode the
number of features of the dataset.

These configurations create pipeline runs that take hours to complete. To
mitigate this issue the number of feature of the dataset are trimmed down to
the most important ones.

Signed-off-by: Dimitris Poulopoulos <dimpo@arrikto.com>